moved IMessageItem to types folder
