+++ src/components/pwc-chat/pwc-chat.tsx
@@ -1,5 +1,5 @@
 import { Component, Event, Prop, h, EventEmitter, Listen } from "@stencil/core";
-import { IMessageItem } from "./IMessageItem";
+import { IMessageItem } from "../types/IMessageItem";
 
 @Component({
   tag: "pwc-chat",
